Adding a custom assembly resolver
