+++ src/PowerShell/PartnerCenter.psd1
@@ -3,7 +3,7 @@
 #
 # Generated by: Microsoft Corporation
 #
-# Generated on: 05/07/2020
+# Generated on: 06/26/2020
 #
 
 @{
@@ -11,7 +11,7 @@
     RootModule = 'PartnerCenter.psm1'
 
     # Version number of this module.
-    ModuleVersion = '3.0.9'
+    ModuleVersion = '3.0.10'
 
     # Supported PSEditions
     CompatiblePSEditions = 'Core', 'Desktop'
