+++ src/PowerShell/PartnerCenter.psm1
@@ -3,7 +3,7 @@
 #
 # Generated by: Microsoft Corporation
 #
-# Generated on: 05/07/2020
+# Generated on: 06/26/2020
 #
 
 $PSDefaultParameterValues.Clear()
