+++ src/PowerShell/PowerShell.csproj
@@ -18,7 +18,7 @@
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
     <OutputPath>$(RepoArtifacts)$(Configuration)\</OutputPath>
     <PublishDir>$(OutputPath)</PublishDir>
-    <Version>3.0.9</Version>
+    <Version>3.0.10</Version>
     <LangVersion>latest</LangVersion>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
   </PropertyGroup>
@@ -31,8 +31,9 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="2.2.0" />
+    <PackageReference Include="Microsoft.Graph.Beta" Version="0.19.0-preview" />
     <PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" Version="2.6.0-preview" />
-    <PackageReference Include="Microsoft.IdentityModel.JsonWebTokens" Version="5.6.0" />
+    <PackageReference Include="Microsoft.IdentityModel.JsonWebTokens" Version="6.6.0" />
     <PackageReference Include="PartnerCenter.DotNet" Version="1.15.6" />
     <PackageReference Include="PowerShellStandard.Library" Version="5.1.0" PrivateAssets="All" />
   </ItemGroup>
@@ -62,6 +63,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PreLoadAssemblies Include="$(RepoSrc)lib\NetFxPreloadAssemblies\*.dll" />
     <PreLoadAssemblies Include="$(RepoSrc)lib\System.Security.Cryptography.ProtectedData.dll" />
   </ItemGroup>
 
@@ -70,16 +72,7 @@
     <ProjectReference Include="..\Subscription\Subscription.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Reference Include="Microsoft.Graph.Beta">
-      <HintPath>..\lib\Microsoft.Graph.Beta.dll</HintPath>
-    </Reference>
-    <Reference Include="Microsoft.Graph.Core">
-      <HintPath>..\lib\Microsoft.Graph.Core.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
   <Target Name="CopyFiles" AfterTargets="Build">
-    <Copy SourceFiles="@(PreLoadAssemblies)" DestinationFolder="$(TargetDir)" />
+    <Copy SourceFiles="@(PreLoadAssemblies)" DestinationFolder="$(TargetDir)\PreloadAssemblies" />
   </Target>
 </Project>
